style : Drawer 열려도 헤더 보이도록 고정
